requirements.txt
absl-py==2.1.0
alembic==1.15.1
annotated-types==0.7.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
category_encoders==2.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
dacite==1.9.2
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_colorscales==0.0.4
dash_daq==0.5.0
databricks-sdk==0.46.0
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
docker==7.1.0
docopt==0.6.2
dtale==3.16.1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filterpy==1.4.5
Flask==3.0.3
Flask-Compress==1.17
flatbuffers==25.2.10
fonttools==4.56.0
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.70.0
h5py==3.13.0
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kaleido==0.2.1
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.44.0
lz4==4.4.3
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.1.2
ml-dtypes==0.4.1
mlflow==2.20.4
mlflow-skinny==2.20.4
multimethod==1.12
namex==0.0.8
narwhals==1.28.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
openpyxl==3.1.5
opentelemetry-api==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.15
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
particles==0.4
patsy==1.0.1
phik==0.12.4
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycaret==3.3.2
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.19.1
pykalman==0.10.1
pyod==2.0.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyWavelets==1.8.0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
ruptures==1.1.9
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
smmap==5.0.2
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.5.3
squarify==0.4.4
stack-data==0.6.3
statsmodels==0.14.4
strsimpy==0.2.1
sweetviz==2.3.1
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
-e git+https://github.com/SpartanFOU/Thesis-Project.git@686863b359ad34dc12a30c9341c003f42fd03c74#egg=timeseries_forecasting_project
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typeguard==4.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
visions==0.7.6
waitress==3.0.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.2
xarray==2025.1.2
xlrd==2.0.1
xxhash==3.5.0
yarg==0.1.9
ydata-profiling==4.12.2
yellowbrick==1.5
zipp==3.21.0
zstandard==0.23.0
